face-detection/requirements.txt
face-recognition
paho-mqtt
python-dotenv
opencv-python
numpy